+++ Red Panda Bazaar Code/Features/Bank/BankData.cs
@@ -2,8 +2,8 @@ namespace Red_Panda_Bazaar_Code.Features.Bank;
 
 public class BankSaveData
 {
-    public int CheckingBalance { get; set; }
-    public int InterestEarned { get; set; }
+    public long CheckingBalance { get; set; }
+    public long InterestEarned { get; set; }
     public List<FixedDeposit> FixedDeposits { get; set; } = new();
     public List<LoanAccount> Loans { get; set; } = new();
     public int LastInterestDay { get; set; }
@@ -11,7 +11,7 @@ public class BankSaveData
 
 public class FixedDeposit
 {
-    public int Amount { get; set; }
+    public long Amount { get; set; }
     public int TermDays { get; set; }
     public int StartDay { get; set; }
     public bool Withdrawn { get; set; }
@@ -20,17 +20,17 @@ public class FixedDeposit
 public class LoanAccount
 {
     public int PlanType { get; set; }
-    public int Principal { get; set; }
+    public long Principal { get; set; }
     public int StartDay { get; set; }
-    public int InterestAccrued { get; set; }
+    public long InterestAccrued { get; set; }
     public bool Repaid { get; set; }
 }
 
 /// <summary>多人同步用的包裹，避免直接暴露内部数据类结构。</summary>
 public class BankSyncData
 {
-    public int CheckingBalance { get; set; }
-    public int InterestEarned { get; set; }
+    public long CheckingBalance { get; set; }
+    public long InterestEarned { get; set; }
     public List<FixedDeposit> FixedDeposits { get; set; } = new();
     public List<LoanAccount> Loans { get; set; } = new();
     public int LastInterestDay { get; set; }
